requirements/base.txt
black==23.3.0
Django==3.2.19
django-classy-tags==4.0.0
django-cms==3.11.3
django-colorfield==0.9.0
django-compressor==4.4
django-db-logger==0.1.12
django-environ==0.10.0
django-filer==3.0.3
django-meta==2.3.0 # new version gives this error AttributeError: module 'meta.settings' has no attribute 'FB_APPID'
django-mptt==0.14.0
django-parler==2.3
django-recaptcha2==1.4.1
django-recaptcha3==0.4.0
django-redis==5.3.0
django-robots==5.0
django-sekizai==4.1.0
django-solo==2.1.0
django-treebeard==4.4
djangocms-admin-style==3.2.4
djangocms-blog==1.2.3
djangocms-bootstrap4==3.0.0
djangocms-cascade==2.3.14
djangocms-file==3.0.0
djangocms-googlemap==2.1.0
djangocms-icon==2.0.0
djangocms-link==3.1.1
djangocms-picture==4.0.0
djangocms-snippet==3.1.0
djangocms-style==3.0.0
djangocms-text-ckeditor==5.1.2
djangocms-video==3.0.0
djangorestframework==3.14.0
easy_thumbnails==2.8.5
googlemaps==4.10.0
html5lib==1.1
Pillow==9.5.0
pre-commit==3.3.3
# psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
sentry-sdk==1.25.1
six==1.16.0
uszipcode==1.0.1
uWSGI==2.0.21
python-Levenshtein
beautifulsoup4==4.9.3